Moved Roche radius calculation to utils.py
